@sanderegg review: rename database by repository
