+++ services/catalog/src/simcore_service_catalog/api/rest/_services.py
@@ -24,8 +24,8 @@ from ...clients.director import DirectorApi
 from ...models.services_db import ServiceAccessRightsAtDB, ServiceMetaDataDBGet
 from ...repository.groups import GroupsRepository
 from ...repository.services import ServicesRepository
-from .._dependencies.database import get_repository
 from .._dependencies.director import get_director_api
+from .._dependencies.repository import get_repository
 from .._dependencies.services import get_service_from_manifest
 
 _logger = logging.getLogger(__name__)
